Remove Num's dependency on the Calculator Configuration
The Num backend is now entirely independent from the DiscoDNC. This
completes the previous work that removed the DsicoDNC-specific Epsilon
stored in each Num implementation. As a result, the DiscoDNC and the Num
backend it uses are separated. Num is thus ready to be moved into its
own repository.

Also, renamed NumImpl (implementation etc.) to NumBackend.