Prevent adding duplicate arrival bounds; readded old behavior
If this analysis configuration option is enabled and there are multiple
arrival bounding alternatives checked in parallel, duplicate results
will be removed.
The literature provides a better option that always results in a single
arrival bound per call of the dispatcher: convolve all alternative
results into a single one.
The former is implemented in the current HEAD of master and v2.5.
However, an intermediate version of the DiscoDNC somehow had a
preliminary implementation of the second alternative. During rebase, I
retained the second, the better alternative. Yet, the implementation is
needs some more additions (e.g., proper naming throughout). Thus, I
reverted to the old code (HEAD) and postpone a proper implementation of
the convolution of alternatives.