Arrival burst reduction; better check for applicability
The current implementation only works for affine curves, i.e., for token
buckets. This setting is now checked before anything else is computed.
Before, we later checked if the curve has only two segments. I.e., this
imposes a restriction as we cannot reduce the burst if a token bucket
was created with pwaffine curves. I do not consider this a problem as
the current code is intended to reproduce the research paper's results.
In it, we use affine curves only. A more generic implementation is
needed here.