Clarified naming for max Service Curve flags and variables
Renamed to a role-based, descriptive name instead of the commonly used
Greek letter gamma that is more commonly used for token-bucket arrival
constraints.
gamma was generally renamed to max_sc or maxSc (depending on the
context, may also be enforce instead of use) and extra_gamma was renamed
to max_sc_output_rate or some variation of it according to the max_sc
renaming.
Servers' getServiceCurve and getGamma were merged, using the previous
getGamma semantics to return a delta_0 curve if the maxSc flag is false.
Thus, a mehtod getStoredMaxSc was added to allow access to the actually
stored curve. The Output bound code was adapted to use the unconditional
access to the stored maximum service curve (rate) if the enforcement
flag is set to GLOBALLY_ON. That should result in expected behavior (was
probably buggy before)

Fixed a bug in Server's setMaxServiceCurve: given flags were ignored