Clarify capabilities of code by separation in according packages
Before version 2.2.5, the deconvolution implementation was falsely
assumed to be capable of deconvolving any piecewise linear curves. The
new implementation provided with v2.2.5 restricts the implementation to
convex or (almost) concave, (almost) continuous functions defined by
 - the maximum over multiple rate-latency curves (mRL, convex
functions),
 - the minimum over multiple token-bucket curves (mTB, almost
concave/continuous functions as f(0)=0),
 - the delayed infinite burst functions \delta_{T} used in the min/max
above.
We now abbreviate this class by ConPwAffine and adapted the provided
MinPlus backend accordingly.

Moreover, computation of bounds was in a package called pwaffine. While
this is true for most of the bound implementations in it, the output
bound is based on the deconvolution and thus inherits its restrictions.
Therefore, all the suffixes to the bound package were dropped to not
indicate any restriction or generality.