Move the checks configuration from top level to lower levels
There are several checks regarding curve shapes that may be done either
by min-plus-algebraic operations, bounding operations, or curve
operations. All of them are specific to the implementation of those yet
they were enabled on the highest possible level in the code, i.e., in
the Calculator. Also, a single configuration flag might have affected
several of the code areas (modules). Therefore, the configurations have
now been moved to the most suitable packages for them such that they are
close to the modules using them and such that they can be swapped out
with those modules if necessary.