Add and enable the arrival bounds cache
This is basically the cache used in the SIGMETRICS 2017 publication but
with one generalization. SIGMETRICS assumed them to be token bucket
curves with equal long-term rates. When "convolving" such alternative
arrival bounds, that code just picked the curve with smallest burst to
be stored. This version is more generic as actually executes a
convolution of all alternatives.