Make skipping the server backlog cap verbose
This is theory presented in the paper MMB 2016 "Improving Cross-Traffic
Bounds in Feed-Forward Networks – There is a Job for Everyone". But like
the paper's evaluation, the DNC implementation is for token bucket
arrivals only. Even if the according AnalysisConfig flag is set, it is
only executed with the AlgDncBackend_DNC_Affine.DISCO_AFFINE backend.
Otherwise, it was silently skipped. This skipping now triggers a
terminal output.