Analysis Interface extended and implemented by AbstractAnalysis
This way, any variable of type Analysis can hold any of the offered
tandem analyses (was possible before) AND we have access to the bounds
etc. The second part was not possible before as the methods to access
them haad common implementations in AbstractAnalysis but were not
defined in the Analysis Interface.
Now AbstractAnalysis implements Analysis,
{Pmoo,SeparateFlowm...}Analysis extends AbstractAnalysis and thus
implicitly implements Analysis. I.e., I removed the explicit,
superfluous implements statement from the classes.